refactor(api): :recycle: modularize code into package
