+++ apps/backend/project.json
@@ -9,7 +9,7 @@
     "build": {
       "executor": "nx:run-commands",
       "options": {
-        "command": "bun build apps/backend/src/index.ts"
+        "command": "bun build --target bun --outdir ./apps/backend/dist apps/backend/src/index.ts"
       }
     },
     "dev": {
@@ -21,7 +21,7 @@
     "start": {
       "executor": "nx:run-commands",
       "options": {
-        "command": "cross-env NODE_ENV=production bun apps/backend/src/index.ts"
+        "command": "cross-env NODE_ENV=production bun apps/backend/dist/index.js"
       }
     }
   }
