+++ libs/api/v1/README.md
@@ -1,0 +1,3 @@
+# api-v1
+
+This library was generated with [Nx](https://nx.dev).
