+++ libs/api/v1/project.json
@@ -1,0 +1,9 @@
+{
+  "name": "api-v1",
+  "$schema": "../../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "libs/api/v1/src",
+  "projectType": "library",
+  "tags": [],
+  "// targets": "to see all targets run: nx show project api-v1 --web",
+  "targets": {}
+}
