+++ libs/api/v1/src/index.ts
@@ -1,0 +1,1 @@
+export * from './lib/api-v1'
