+++ tsconfig.base.json
@@ -16,6 +16,7 @@
     "baseUrl": ".",
     "paths": {
       "@nmit-coursition/ai": ["libs/ai/src/index.ts"],
+      "@nmit-coursition/api/v1": ["libs/api/v1/src/index.ts"],
       "@nmit-coursition/design-system": ["libs/ui/design-system/src/index.ts"],
       "@nmit-coursition/email": ["libs/email/src/index.ts"],
       "@nmit-coursition/payments": ["libs/payments/src/index.ts"],
